kindle-to-pdf/requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
# 以下のライブラリはAI OCRを使用する場合のみ必要
# openai>=1.0.0
# google-generativeai>=0.3.0
